Fixed stuff ready for release
Updated readme and help page to include information about Invoke-PMUninstall.
Updated changelong & manifest to 0.1.2.
Updated help pages to reference uninstall-scritpblocks.
Updated build script to publish version number, used in release pipeline.

+++ ProgramManager/functions/Invoke-PMUninstall.ps1
@@ -125,9 +125,13 @@
 				
 			}
 			
-			# Delete the package files
-			Write-Verbose "Deleting package files"
-			Remove-Item -Path "$($package.InstallDirectory)\$($package.Name)" -Recurse -Force
+			if ($PSCmdlet.ShouldProcess("Package:{$name} files", "Delete from the installation directory")) {
+				
+				# Delete the package files
+				Write-Verbose "Deleting package files"
+				Remove-Item -Path "$($package.InstallDirectory)\$($package.Name)" -Recurse -Force
+				
+			}
 			
 		}elseif ($package.Type -eq "ChocolateyPackage") {
